# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Snappfood API",
        "description": "API documentation for the Snappfood.",
        "contact": {
            "email": "farshidboroomand@gmail.com.com"
        },
        "version": "V1"
    },
    "paths": {
...
        "/api/v1/users/{user_id}/profiles/{profile_id}": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user profile",
                "description": "Allows updating the user's sheba number and national ID.",
                "operationId": "updateUserProfile",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "UUID of the user",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "2bb3ded7-00be-4763-8016-9d072ba374d3"
                        }
                    },
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "UUID of the user profile",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a4b73c3c-5d3b-4f6b-9f58-0935b65430c2"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sheba_number": {
                                        "description": "Iranian sheba number (must be 24 characters, start with IR)",
                                        "type": "string",
                                        "example": "IR123456789012345678901234"
                                    },
                                    "national_id": {
                                        "description": "Iranian national ID (exactly 10 digits)",
                                        "type": "string",
                                        "example": "1234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "profileId": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "a4b73c3c-5d3b-4f6b-9f58-0935b65430c2"
                                                },
                                                "userId": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "2bb3ded7-00be-4763-8016-9d072ba374d3"
                                                },
                                                "shebaNumber": {
                                                    "type": "string",
                                                    "example": "IR123456789012345678901234"
                                                },
                                                "nationalId": {
                                                    "type": "string",
                                                    "example": "1234567890"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found"
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "sheba_number": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The sheba number format is invalid."
                                                    }
                                                },
                                                "national_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The national id must be 10 digits."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/wallets/{wallet_id}": {
            "get": {
                "tags": [
                    "Wallets"
                ],
                "summary": "Fetch a user's wallet by wallet ID",
                "description": "Returns wallet details along with user information.",
                "operationId": "getUserWallet",
                "parameters": [
                    {
                        "name": "wallet_id",
                        "in": "path",
                        "description": "UUID of the wallet",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "9e3342a5-621f-4d03-9a96-cd113fbba05f"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wallet fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "walletId": {
                                                    "type": "string",
                                                    "example": "9e3342a5-621f-4d03-9a96-cd113fbba05f"
                                                },
                                                "balance": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1000000
                                                },
                                                "availableBalance": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 800000
                                                },
                                                "blockedAmount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 200000
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "IRT"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "userId": {
                                                            "type": "string",
                                                            "example": "f0576399-38fc-4846-bc16-a09d16a44437"
                                                        },
                                                        "firstName": {
                                                            "type": "string",
                                                            "example": "Noah"
                                                        },
                                                        "lastName": {
                                                            "type": "string",
                                                            "example": "Lakin"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "noah.lakin@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Wallet not found"
                    },
                    "500": {
                        "description": "Server error while retrieving wallet"
                    }
                }
            }
        },
...
    },
...
}